backend/pyproject.toml
[project]
name = "tfl-explorer-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.2.0",
    "agent-framework>=1.0.0rc1",
    "neo4j-agent-memory[openai,microsoft-agent]",
    "neo4j>=5.26.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "testcontainers[neo4j]>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
markers = [
    "unit: fast isolated unit tests",
    "integration: tests requiring Neo4j (docker or env)",
    "e2e: end-to-end smoke tests against running server",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.uv]
prerelease = "allow"
